Optimize Day 11
* Use a pattern where people appear and disappear until they are in a
  position that can't change (because one person has only empty
  neighbours and no-one will seat there since there is a neighbour
  touching it) to simplify the problem to a boolean state rather than a
  ternary state.
* Cache the neighbours from any position and the visible seats for part
  2. The latter is important as it's the only way mapping values to
  bools works since Floor cells are the only visited cells in the
  initial state and we must know which ones they are to compute the
  visible positions
* Keep a smaller cache of visited rows and columns to allow skipping an
  entire row or column at once when all positions have already been
  visited and won't change anymore.

This is not the end of what's possible. Currently, the code is a bit
messy and could be refactored.
The Vec<bool> to skip lines and columns is a rudimentary.
It would be smarter to keep start and end indices for regions to skip
so we can save ourselves a little bit more work.

The runtime was too high for convenient benchmarking, but for anecdotal
evidence, before:
```
AOC 2020
Day 11 - Part 1 : 2427
	generator: 39.017µs,
	runner: 237.225871ms

Day 11 - Part 2 : 2199
	generator: 47.942µs,
	runner: 230.390933ms
```

after:
```
AOC 2020
Day 11 - Part 1 : 2427
	generator: 351ns,
	runner: 8.546524ms

Day 11 - Part 2 : 2199
	generator: 174ns,
	runner: 12.758047ms
```