Day 20: stop doing twice the necessary work
Flipping the pieces around the x axis and the y axis is unnecessary as
we will hit the same position after rotating the piece and flipping it.

I managed to convince myself otherwise on the day, but it was a
miscalculation.

An easy way to convince oneself is to think of an actual jigsaw puzzle:
whether you flip a tile around the x or the y axis, there are only two
flipped states: face up or face down.